requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
fastapi==0.115.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytz==2024.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
tortoise-orm==0.21.7
typing_extensions==4.12.2
uvicorn==0.32.0
